requirements.txt
aiogram==3.3.0
openpyxl
python-docx